+++ DefectListDomain/Models/MapBomItemToRouteChart.cs
@@ -9,7 +9,7 @@ namespace DefectListDomain.Models
         public string RouteChart_Number { get; set; }
         public string Detal { get; set; }
         public int ProductId { get; set; }
-        public float QtyLaunched { get; set; }
+        public decimal QtyLaunched { get; set; }
         public DateTime CreateDate { get; set; }
         public string CreatedBy { get; set; }
         public string CreatedByName { get; set; }
